refactor: change naming fav to favs
