--- src/favs/dto/create-fav.dto.ts
@@ -1,10 +1,0 @@
-import { IsUUID, IsEnum } from 'class-validator';
-import { FavTypes } from '../entities/fav.entity';
-
-export class CreateFavDto {
-  @IsUUID()
-  id: string;
-
-  @IsEnum(FavTypes)
-  type: FavTypes;
-}
