+++ src/favs/dto/update-favs.dto.ts
@@ -1,4 +1,4 @@
 import { PartialType } from '@nestjs/swagger';
-import { CreateFavDto } from './create-fav.dto';
+import { CreateFavDto } from './create-favs.dto';
 
 export class UpdateFavDto extends PartialType(CreateFavDto) {}
