+++ src/utils/DB/entities/DBFav.ts
@@ -1,10 +1,10 @@
-import { CreateFavDto } from 'src/favs/dto/create-fav.dto';
-import { UpdateFavDto } from 'src/favs/dto/update-fav.dto';
-import { FavEntity } from 'src/favs/entities/fav.entity';
+import { CreateFavDto } from 'src/favs/dto/create-favs.dto';
+import { UpdateFavDto } from 'src/favs/dto/update-favs.dto';
+import { FavsEntity } from 'src/favs/entities/favs.entity';
 import DBEntity from './DBEntity';
 
 export default class DBFav extends DBEntity<
-  FavEntity,
+  FavsEntity,
   UpdateFavDto,
   CreateFavDto
 > {
